Refactor formatting stuff into its own module
